Made some fixes in the scenes like: - player was too fast - put some collider from non-trigger to trigger - resized wall colliders - fixed moving with controller and falling through the floor - teleport ray not shown due to layer interferences
PROBLEM WHEN MOVING/ROTATING WITH CONTROLLERS: We have that an Interactable MUST have a collider and a rigibody, but appartently a TRIGGER collider will act as NON-trigger if it has a rigidbody attached.
- Removing the rigidbody from the Interactable stops the logic of interacts-with, so we must keep the rigidbody attached.
 - That's a big problem because we have some big trigger colliders in front of the paintings, causing the player to "collide" with an invisible wall.
What I'm going to do RIGHT NOW is to get rid of these big colliders in front of the paintings, in the hope we would not need to use them in the future. I keep the colliders as big as the paintings.
AMEND: There was still the problem of the player not able to get close to walls/paintings. I disabled the character controller of the player. Now we would get stuck with anything, but theorically speaking we can go through walls.